* `PickerInputBaseProps` interface was moved from @epam/uui-components to '@epam/uui-core' package * [PickerInput/FiltersPickerInput]: always show picker footer if picker has some selection * [FiltersPanel]: add 'minCharsToSearch' and 'renderFooter' props for picker filter configuration
